+++ operator/tests/tutils/ais_cluster.go
@@ -44,7 +44,6 @@ type (
 		InitImage                 string
 		LogSidecarImage           string
 		DisableTargetAntiAffinity bool
-		EnableExternalLB          bool
 		ProxyExternalAccess       bool
 		TargetExternalAccess      bool
 		EnableAdminClient         bool
@@ -191,8 +190,7 @@ func newAISClusterCR(args *ClusterSpecArgs, mounts []aisv1.Mount) *aisv1.AIStore
 		LogSidecar: &aisv1.LogSidecarSpec{
 			Image: args.LogSidecarImage,
 		},
-		StateStorage:     args.StateStorage,
-		EnableExternalLB: args.EnableExternalLB,
+		StateStorage: args.StateStorage,
 		ProxySpec: aisv1.DaemonSpec{
 			ServiceSpec: aisv1.ServiceSpec{
 				ServicePort:      intstr.FromInt32(51080),
@@ -215,14 +213,14 @@ func newAISClusterCR(args *ClusterSpecArgs, mounts []aisv1.Mount) *aisv1.AIStore
 			DisablePodAntiAffinity: &args.DisableTargetAntiAffinity,
 		},
 	}
-	if args.EnableExternalLB || args.ProxyExternalAccess {
+	if args.ProxyExternalAccess {
 		spec.ProxySpec.ExternalAccess = &aisv1.ExternalAccessSpec{}
 	}
-	if args.EnableExternalLB || args.TargetExternalAccess {
+	if args.TargetExternalAccess {
 		spec.TargetSpec.ExternalAccess = &aisv1.ExternalAccessSpec{}
 	}
 	// If not using an LB, use the host port to provide external access
-	if !args.EnableExternalLB && !args.ProxyExternalAccess && !args.TargetExternalAccess {
+	if !args.ProxyExternalAccess && !args.TargetExternalAccess {
 		spec.APIMode = aisapc.Ptr(args.APIMode)
 		spec.ProxySpec.HostPort = aisapc.Ptr(int32(51080))
 		spec.TargetSpec.HostPort = aisapc.Ptr(int32(51081))